+++ app/Main.hs
@@ -31,13 +31,14 @@ main = withGetOpt "[options] [input-file]" opts $ \r args -> do
   case toCmd (#input @= args <: r) of
     PrintVersion    -> B.putStr $ fromString (showVersion version <> "\n")
     CallMeAPI opts' -> run callMeAPI opts'
-    PostStory opts' -> run (postStory $ getPath opts') opts'
+    PostStory opts' -> run (postStory (getPath opts') (opts' ^. #title)) opts'
   where
     loadEnvFileIfExist conf =
       whenM (and <$> mapM doesFileExist (configPath conf)) (void $ loadFile conf)
     opts = #version @= versionOpt
         <: #verbose @= verboseOpt
         <: #me      @= meOpt
+        <: #title   @= titleOpt
         <: nil
     getPath opts' =
       fromMaybe (error "please input filepath") $ listToMaybe (opts' ^. #input)