+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased changes
 
+- Feat: `--title` option (explicit specify post title)
+
 ## 0.1.0.0
 
 - Feat: Post Markdown story to Medium